🚚(frontend) move ContractFrame as component
ContractFrame is a generic comopnent that can be called from anywhere
