+++ src/frontend/js/components/ContractFrame/_styles.scss
