+++ src/frontend/js/components/ContractFrame/iframe-manager.js
