+++ src/frontend/js/components/ContractFrame/index.tsx
