+++ src/frontend/scss/components/_index.scss
@@ -1,4 +1,5 @@
 @import '../../js/components/AddressesManagement/styles';
+@import '../../js/components/ContractFrame/styles';
 @import '../../js/components/TeacherDashboardCourseList/styles';
 @import '../../js/components/Modal/styles';
 @import '../../js/components/PaymentButton/styles';
@@ -26,7 +27,6 @@
 @import '../../js/widgets/Dashboard/components/DashboardBox/styles';
 @import '../../js/widgets/Dashboard/components/DashboardCard/styles';
 @import '../../js/widgets/Dashboard/components/DashboardItem/Certificate/styles';
-@import '../../js/widgets/Dashboard/components/DashboardItem/Order/ContractFrame/styles';
 @import '../../js/widgets/Dashboard/components/DashboardItem/Order/styles';
 @import '../../js/widgets/Dashboard/components/DashboardItem/styles';
 @import '../../js/widgets/Dashboard/components/DashboardLayout/styles';
